# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    API สำหรับระบบจัดการงานและทรัพยากรองค์กร (Task & Resource Management System)
    
    ## Authentication
    API ใช้ JWT Bearer Token สำหรับการยืนยันตัวตน
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Login endpoint: 5 requests/minute per IP
    - Other endpoints: ไม่จำกัด (ในเวอร์ชันปัจจุบัน)
    
    ## Roles
    - **ADMIN** - ผู้ดูแลระบบ
    - **CUSTOMER_SERVICE** - ฝ่ายบริการลูกค้า
    - **FINANCE_LEADER** - หัวหน้าฝ่ายการเงิน
    - **FINANCE** - เจ้าหน้าที่การเงิน
    - **SALES_LEADER** - หัวหน้าฝ่ายขาย
    - **SALES** - พนักงานขาย
    - **HEAD_TECH** - หัวหน้าแผนกช่าง
    - **LEADER** - หัวหน้าทีม
    - **TECH** - ช่าง
    
  version: 1.0.0
  contact:
    name: TaskFlow Support
    email: support@taskflow.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://taskflow.example.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token ที่ได้จาก /auth/login
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
        error:
          type: string
...
    LeaveType:
      type: string
      enum:
        - SICK
        - PERSONAL
        - VACATION
        - BIRTHDAY
        - OTHER
    LeaveStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
        - CANCELLED
    LeaveDurationType:
      type: string
      enum:
        - FULL_DAY
        - HALF_DAY
        - TIME_BASED
...
    CreateLeaveRequest:
      type: object
      required:
        - type
        - startDate
        - endDate
      properties:
        type:
          $ref: '#/components/schemas/LeaveType'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
        durationType:
          $ref: '#/components/schemas/LeaveDurationType'
        halfDayPeriod:
          type: string
          enum: [MORNING, AFTERNOON]
        startTime:
          type: string
        endTime:
          type: string
  responses:
    Unauthorized:
      description: ไม่ได้ยืนยันตัวตน
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: กรุณาเข้าสู่ระบบ
    Forbidden:
      description: ไม่มีสิทธิ์เข้าถึง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: ไม่มีสิทธิ์เข้าถึง
    BadRequest:
      description: ข้อมูลไม่ถูกต้อง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: กรุณากรอกข้อมูลให้ครบถ้วน
    NotFound:
      description: ไม่พบข้อมูล
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: ไม่พบข้อมูล
...
paths:
...
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: ดึงรายละเอียดงาน
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: รายละเอียดงาน
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Tasks]
      summary: อัปเดตงาน
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: อัปเดตสำเร็จ
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: ลบงาน
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ลบสำเร็จ
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== Cars ====================
...
  /leaves:
    get:
      tags: [Leaves]
      summary: ดึงรายการลา
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/LeaveStatus'
        - name: userId
          in: query
          schema:
            type: string
        - name: pendingApproval
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: รายการลา
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Leaves]
      summary: ส่งคำขอลา
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeaveRequest'
      responses:
        '200':
          description: ส่งคำขอลาสำเร็จ
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /notifications:
    get:
      tags: [Notifications]
      summary: ดึงการแจ้งเตือน
      security:
        - bearerAuth: []
      responses:
        '200':
          description: รายการแจ้งเตือน
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Notifications]
      summary: อัปเดตสถานะการอ่าน
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isRead:
                  type: boolean
      responses:
        '200':
          description: อัปเดตสำเร็จ
        '401':
          $ref: '#/components/responses/Unauthorized'